requirements.txt
numpy>=2.0
pandas>=2.2
pydantic>=2.0
PyYAML>=6.0
pytest>=9.0
